requirements.txt
python-telegram-bot[job-queue]==21.10
supabase>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
